requirements.txt
gelsight==3.0
imageio==2.31.1
ipykernel==6.23.3
ipython==8.12.3
ipywidgets==8.0.6
joblib==1.2.0
matplotlib==3.7.1
matplotlib-inline==0.1.6
numpy==1.24.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
opencv-python==4.9.0.80
Pillow==9.5.0
pip==23.1.2
PyAudio==0.2.13
pyrealsense2==2.54.1.5217
scikit-image==0.21.0
scikit-learn==1.2.2
scipy==1.11.0
setuptools==67.8.0
sounddevice==0.4.6
soundfile==0.12.1
tensorboard==2.13.0
tensorboard-data-server==0.7.1
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
tqdm==4.65.0
xArm-Python-SDK==1.13.8
